feature(#01): Create initial project structure with default configuration
Create three main folders for each part of the project with added Dockerfiles and docker-compose.
Create a NextJS app with prettier enabled and its configuration for building.
Create a Flask application, which will be used for API and for launching the scraper app.
Create a scrapper folder where the Python subpackage will be created.
Add everything into README.md.

This commit message is also used as an example for following commit messages.